#157: reactions as feature
