+++ src/webui/components/module.ts
@@ -6,6 +6,7 @@ export { default as ApplicationSidebar } from './application/sidebar/Sidebar';
 export { default as ErrorBoundary } from './common/ErrorBoundary';
 export { default as Loading } from './common/Loading';
 export { default as LogoutPanel } from './common/LogoutPanel';
+export { default as OrderAndAddRow } from './common/OrderAndAddRow';
 export { default as OrderAndSearchRow } from './common/OrderAndSearchRow';
 export { default as OrderRow } from './common/OrderRow';
 export { default as NotificationPanelList } from './notification/PanelList';
@@ -13,3 +14,4 @@ export { default as PostPanelGrid } from './post/PanelGrid';
 export { default as PostPanelList } from './post/PanelList';
 export { default as RelationPanelList } from './relation/PanelList';
 export { default as RelationProfile } from './relation/Profile';
+
